fix clirr check
