+++ google-cloud-datastore/src/main/java/com/google/cloud/datastore/spi/v1/RpcUtils.java
@@ -19,26 +19,21 @@ package com.google.cloud.datastore.spi.v1;
 import static com.google.api.gax.rpc.StatusCode.Code.ABORTED;
 import static com.google.api.gax.rpc.StatusCode.Code.DEADLINE_EXCEEDED;
 import static com.google.api.gax.rpc.StatusCode.Code.UNAVAILABLE;
-import static com.google.common.truth.Truth.assertThat;
 
+import com.google.api.core.ApiFunction;
+import com.google.api.core.InternalApi;
 import com.google.api.gax.rpc.UnaryCallSettings;
 import com.google.cloud.datastore.DatastoreOptions;
-import com.google.common.collect.Sets;
-import org.junit.Test;
 
-public class DatastoreRpcTest {
-
-  @Test
-  public void testRetrySettingSetter() {
-    DatastoreOptions datastoreOptions =
-        DatastoreOptions.newBuilder().setProjectId("project-id").build();
-    UnaryCallSettings.Builder<Object, Object> builder =
-        UnaryCallSettings.newUnaryCallSettingsBuilder();
-
-    DatastoreRpc.retrySettingSetter(datastoreOptions).apply(builder);
-
-    assertThat(builder.getRetryableCodes())
-        .isEqualTo(Sets.newHashSet(ABORTED, UNAVAILABLE, DEADLINE_EXCEEDED));
-    assertThat(builder.getRetrySettings()).isEqualTo(datastoreOptions.getRetrySettings());
+@InternalApi
+public class RpcUtils {
+  @InternalApi
+  static ApiFunction<UnaryCallSettings.Builder<?, ?>, Void> retrySettingSetter(
+      DatastoreOptions datastoreOptions) {
+    return builder -> {
+      builder.setRetryableCodes(ABORTED, DEADLINE_EXCEEDED, UNAVAILABLE);
+      builder.setRetrySettings(datastoreOptions.getRetrySettings());
+      return null;
+    };
   }
 }
